Move render into service directory
