+++ tests/Service/RenderTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Jstewmc\Rtf;
+namespace Jstewmc\Rtf\Service;
+
+use Jstewmc\Rtf\Element;
 
 class RendererTest extends \PHPUnit\Framework\TestCase
 {
